Update README.md
Digital Circus for it's story as well as one of the big indie animations making headlines, 
Hazbin Hotel again another one of these big indie animations coming into Hollywood and reaching Amazon Prime. 
Overlord, the anime is peak isekai. 
Half-life series is a start of the generation in 2012 that use Gmod for basically everything in animation. 
Last Mandela Catalogue for being the pivot for analog horror